pyproject.toml
[project]
name = "video-search-desktop"
version = "1.0.0"
description = "Búsqueda visual en videos de seguridad por lenguaje natural — 100% local"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [
    { name = "Video Search Team" },
]

dependencies = [
    # ── GUI ──
    "PySide6>=6.8",

    # ── Computer Vision ──
    "opencv-python-headless>=4.10",
    "ultralytics>=8.3",
    "Pillow>=11.0",
    "numpy>=1.26",

    # ── AI / ML ──
    "torch>=2.5",
    "torchvision>=0.20",
    "sentence-transformers>=3.3",
    "transformers>=4.46",
    "bitsandbytes>=0.44",
    "accelerate>=1.2",
    "qwen-vl-utils>=0.0.8",

    # ── Database ──
    "chromadb>=1.5.7",

    # ── Data Validation ──
    "pydantic>=2.9",
    "pydantic-settings>=2.6",

    # ── Logging ──
    "loguru>=0.7",

    # ── GPU Monitoring ──
    "nvidia-ml-py>=12.560",

    # ── Config ──
    "python-dotenv>=1.0",
]

[project.scripts]
video-search = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "models", "ui"]